Drop some code related to test context from OpenApiDocumentService
This code was added to reset some context for tests but it's in some
production code.
While it could look totally benign, it's actually not because if we
haven't initialized the OpenApiDocumentService CDI bean during runtime
(which could happen really), we will initialize it when we shut down the
application, making the shutdown slower than we want (which might or
might not be a problem depending on your use case).

I reached out to the original author and we don't really know why this
code was added and it might be useless now.
And Roberto doesn't like system properties :).

I ran the tests locally, let's see what CI has to say.