Allow multiple format and themes in the Config Doc generator
The idea is to allow generating Markdown for easily publishing the
config on GitHub.

The Javadoc -> Markdown part will probably change a lot soon as we need
to support Markdown for the IDEs.

This is the fruit of a collaboration with @Athou (Jérémie Panzer), who
is marked as a co-author.

Co-authored-by: Jérémie Panzer <jeremiepanzer@gmail.com>