Config Doc - Avoid annotations in primitive type name
While this was already correctly handled for declared types, primitive
types don't have an API to get the name without the annotations being
present (toString() contains the annotations and there is no other way
to get a string representation of the type...).

Fixing it by making sure we get to the type name.

This avoids having things like:
@io.smallrye.config.WithConverter(io.quarkus.runtime.init.InitRuntimeConfig.BooleanConverter.class) boolean
in the doc.
Or Bean Validation constraints when they are used.