Refine datasource tests related to "empty config" and health checks
To be completely sure of what we'll change exactly in further changes.

This is only a refactoring and doesn't change existing test expectations
(though it adds some).

"Empty config" actually means "missing JDBC/reactive URL".
Testing what happens when named datasources are not defined at build
time, in particular, is beside the point.

For health checks, it's just that some configurations were not being
tested, and the next commits may affect the behavior.