Add quarkus.hibernate-orm.database.version-check.enabled
This allows disabling the check on startup if one knows the database
won't be reachable.

It also currently defaults to being disabled when a dialect is set
explicitly (`quarkus.hibernate-orm.dialect=something`),
in order to work around problems we have with correctly
detecting the version on some databases that we don't have tests for
(unsupported ones).